Improve C API error handling
Now `tea_err_throw` actually throws the error and creates an error string with stacktrace, which can be checked and printed in the C API after a `tea_pcall` if `status != TEA_OK`

Next step would have to be to give `tea_pcall` the ability to use a custom error function, allowing us to avoid having a stacktrace attached to the message. Very cool.