Rework maps, disallow `nil` and `nan` keys
It makes implementing the map slightly easier. Now, we can just check if a key is nil to determine where to find a new entry for a new key. Also fixed previous issues with `.delete`. Turns out, it's a bit more tricky than it seems to handle deleted keys, but at least the map allocation is shrink.