Remove string multiply, use `String.repeat`
Thus, operators with different types of values are forbidden from the core Teascript.
It kind of makes sense if you have type check support in the language, but can become confusing otherwise.
I've seen enough Javascript memes to think it's a bad idea.

But hey, if something exists, there will always be somebody hating on it.