Add `pcall` function
While we ponder about the possible way of nicely doing error handling, `pcall` is added to call functions in a protected way. It, like in Lua, returns two values:
- boolean status
- error
Not matter what you think, there has to be a way to capture all kinds of errors thrown by the interpreter, so this is a simple way to handle it at the moment.
Perhaps the error could later be a `Error` object with some useful attributes one can access and use.