Small update to main interpreter loop
Added some nice comments in-between bytecodes, like in `tea_bc.h`.
Now, the main interpreter loop even without computed gotos doesn't really need the `while(true)` we had at the top, because the alternative is a switch-goto. Both create an infinite loop until there are no more instructions to run.
We should strive to keep `tea_vm.c` tight and easy to handle, because it'll really help for a future JIT compiler. It's foolish to clobber the VM with more features.