Drop `++` and `--` operators
Here's another small removal.
It's not a big deal, as `+=` and `-=` have the same effect and avoid subtle side effects.
It also simplifies parsing a tiny bit.
C's handling of `++x` vs. `x++` is very tricky. The implementation is supposed to be fairly simple to write. If I made it, so you can too.