Use bytes for string index/slice, remove `bytelen`
The approach of using UTF-8 codepoints for string indexing, slicing and methods is that it requires string validation, which is very costly.

Unicode is a beast in and of itself.
It's way better to provide a built-in `utf8` module for handling with codepoints.

Therefore strings are left treated as kind of immutable byte arrays.