Argument call spreading
There are 4 kinds of bytecode calls in Teascript VM: `CALL`, `INVOKE`, `SUPER` and `NEW`.

Spread of function arguments turns out to be kind of tricky. Currently, the implementation is basically self-modifying bytecode.

Is it bad? Not sure, but it does work! Surely there's another way to implement it, but for now I think it's good enough?

Well, there are also other parts to fix, like optional function arguments. And later on converting the VM to register based.

Let's see how it goes...